refactor(transit): move display name helpers
