+++ src/domain/transit/name-resolver/get-stop-display-names.ts
@@ -1,5 +1,5 @@
 import type { Stop } from '../../../types/app/transit';
-import type { ResolvedDisplayNames } from '../get-display-names';
+import type { ResolvedDisplayNames } from './get-display-names';
 import { resolveDisplayNamesWithTranslatableText } from '../i18n/resolve-display-names-with-translatable-text';
 
 /**
